// openapi/collector/model/twitterSampleQueryRequest.ts
/**
 * 
 * MPDC Collector API
 *
 * The version of the OpenAPI document: 0.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface TwitterSampleQueryRequest { 
    apiKey: string;
    fetchMultimedia: Array<TwitterSampleQueryRequest.FetchMultimediaEnum>;
    indexInCineast: boolean;
    label: string;
    collectionName: string;
    from: string;
    to: string;
    referencedTweetsDepth?: number;
}
export namespace TwitterSampleQueryRequest {
    export type FetchMultimediaEnum = 'UNKNOWN' | 'ANIMATED_IMAGE' | 'IMAGE' | 'VIDEO';
    export const FetchMultimediaEnum = {
        Unknown: 'UNKNOWN' as FetchMultimediaEnum,
        AnimatedImage: 'ANIMATED_IMAGE' as FetchMultimediaEnum,
        Image: 'IMAGE' as FetchMultimediaEnum,
        Video: 'VIDEO' as FetchMultimediaEnum
    };
}